backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.9
pandas==2.2.0
supabase==2.3.4
python-dotenv==1.0.1
pytest==8.0.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
gspread==5.12.4 